Prevent from admin results from user screens

+++ web/controllers/user/tournament_user_controller.py
@@ -179,6 +179,7 @@ class CheckInUserController(AbstractInputUserController):
             return web_context.error
         return self._user_screen_render(web_context)
 
+
 class IllegalMoveUserController(AbstractInputUserController):
     def _delete_or_add_illegal_move(
             self, request: HTMXRequest,
@@ -286,10 +287,14 @@ class ResultUserController(AbstractInputUserController):
             return AbstractController.redirect_error(
                 request, f'la ronde [{round}] est invalide.')
         if result is None:
+            if not web_context.admin_auth:
+                return AbstractController.redirect_error(
+                    request, f'la suppression de résultat n\'est pas autorisée.')
             with suppress(ValueError):
                 web_context.tournament.delete_result(web_context.board)
         else:
-            if result not in Result.imputable_results():
+            if ((web_context.admin_auth and result not in Result.admin_imputable_results())
+                    or result not in Result.user_imputable_results()):
                 return AbstractController.redirect_error(request, f'Le résultat [{result}] est invalide.')
             web_context.tournament.add_result(web_context.board, Result.from_papi_value(result))
         SessionHandler.set_session_last_result_updated(request, web_context.tournament.id, round, web_context.board.id)
